apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*", "tests*"]

[project]
name = "trust-copilot-api"
version = "0.1.0"
description = "Trust Copilot FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi",
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "alembic",
    "pgvector",
    "openai",
    "boto3",
    "python-docx",
    "openpyxl",
    "pymupdf",
    "pytest",
    "httpx",
    "python-multipart",
    "python-dotenv",
    "pyotp",
    "cryptography",
    "redis",
    "sentry-sdk[fastapi]",
    "prometheus-client",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that require Postgres (skip with -m 'not integration')",
]

[tool.ruff]
line-length = 100
